Extract reusable config methods + start to clean up logging
