+++ packages/release-notes-generator/package.json
@@ -1,6 +1,6 @@
 {
   "name": "release-notes-generator",
-  "version": "1.0.0",
+  "version": "0.0.0",
   "type": "module",
   "main": "./dist/cjs/index.js",
   "module": "./dist/esm/index.js",
@@ -25,7 +25,7 @@
     "test": "vitest"
   },
   "keywords": [],
-  "author": "Nick Larew",
+  "author": "MongoDB, Inc.",
   "license": "ISC",
   "description": "",
   "devDependencies": {
