+++ packages/release-notes-generator/src/llm/classifyChangelog.ts
@@ -2,7 +2,6 @@ import type { AzureOpenAI } from "mongodb-rag-core/openai";
 import { PromisePool } from "@supercharge/promise-pool";
 import { makeClassifyChangelogAudience } from "./classifyChangelogAudience";
 import { makeClassifyChangelogScope } from "./classifyChangelogScope";
-import { iOfN } from "../utils";
 import type { Logger } from "../logger";
 import type { Classification } from "mongodb-rag-core";
 
@@ -82,7 +81,7 @@ export function makeClassifyChangelogs({
         });
         errors.push(error);
       })
-      .process(async (changeDescription, index) => {
+      .process(async (changeDescription) => {
         const summary = await classifyChangelog({
           changelog: changeDescription,
         });
